refactored code to components and containers
